feat: whats new module separated
