+++ packages/hoppscotch-common/locales/en.json
@@ -98,7 +98,9 @@
     "twitter": "Twitter",
     "type_a_command_search": "Type a command or search…",
     "we_use_cookies": "We use cookies",
+    "updated_text": "Hoppscotch has been updated to the version {version} 🎉",
     "whats_new": "What's new?",
+    "see_whats_new": "See what’s new",
     "wiki": "Wiki"
   },
   "auth": {
