+++ packages/hoppscotch-common/src/modules/whats-new.ts
@@ -1,0 +1,8 @@
+import { useWhatsNewDialog } from "~/composables/whats-new"
+import { HoppModule } from "."
+
+export default <HoppModule>{
+  onRootSetup() {
+    useWhatsNewDialog()
+  },
+}
